extract_news/requirements.txt
flask
requests
google-cloud-storage
google-cloud-secret-manager
